api/requirements.txt
fastapi>=0.115.0
mangum>=0.19.0
uvicorn>=0.34.0
boto3>=1.35.0
pydantic>=2.10.0
python-ulid>=3.0.0
python-dateutil>=2.9.0
